backend/requirements.txt
# Django and core dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0

# GraphQL
graphene-django>=3.1.0
graphql-core>=3.2.0

# Database
psycopg2-binary>=2.9.0

# Environment management
python-dotenv>=1.0.0

# Development dependencies (optional)
django-debug-toolbar>=4.0.0